net: avoid unnecessary type conversions
CL generated mechanically with github.com/mdempsky/unconvert.

Reviewed-on: https://go-review.googlesource.com/100035
Run-TryBot: Kunpei Sakai <namusyaka@gmail.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>

+++ src/net/ip.go
@@ -265,17 +265,17 @@ func (ip IP) Mask(mask IPMask) IP {
 // that dst has sufficient length.
 func ubtoa(dst []byte, start int, v byte) int {
 	if v < 10 {
-		dst[start] = byte(v + '0')
+		dst[start] = v + '0'
 		return 1
 	} else if v < 100 {
-		dst[start+1] = byte(v%10 + '0')
-		dst[start] = byte(v/10 + '0')
+		dst[start+1] = v%10 + '0'
+		dst[start] = v/10 + '0'
 		return 2
 	}
 
-	dst[start+2] = byte(v%10 + '0')
-	dst[start+1] = byte((v/10)%10 + '0')
-	dst[start] = byte(v/100 + '0')
+	dst[start+2] = v%10 + '0'
+	dst[start+1] = (v/10)%10 + '0'
+	dst[start] = v/100 + '0'
 	return 3
 }
